# server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fast Ship API"
  version: "0.1.42"
  license:
    name: MIT
  description: |
    Fast Ship 项目管理工具的 HTTP API。覆盖 `server/internal/router/router.go` 注册的全部 /api 路由；
    SPA 静态资源与 NoRoute 回退（`web.go`）不在本文档范围。

    约定：
    - 所有 JSON 响应统一信封 `{code, message, data}`：`code=0` 成功；非 0 为业务错误码，`data` 为 null。
    - 认证方式为 `Authorization: Bearer <token>`；token 为 JWT（登录/注册签发）或 `fsk_` 开头的 API Key，服务端按前缀区分。
    - 部分二进制资源端点（头像、Issue 图片、安装包下载、GitHub 媒体代理）额外接受 `?token=<token>` query 凭证。
    - 除特别注明外，字段级权限写在各字段 description 中：「仅 JWT」表示 API Key 携带该字段会返回 403（40301）；「仅 API Key」表示 JWT 调用返回 403（40303）。
servers:
  - url: /
    description: 与后端服务同源
...
paths:
...
  /api/issues/{iid}/pull-requests:
    post:
      tags: [issues]
      operationId: attachIssuePullRequest
      summary: 关联一个 GitHub PR 到 Issue（幂等）
      description: |
        请求体只传 PR URL（`https://github.com/<owner>/<repo>/pull/<number>`），服务端解析 owner/repo/number 后立即从 GitHub 拉取 title/state/is_draft/author/head/base/merged_at 等字段入库；拉取失败返回错误，不落记录。
        允许跨仓库：URL 可以是任意 GitHub 仓库的 PR，不要求等于项目配置的仓库，不做跨仓库鉴权。项目未配 GitHub token 时以未认证方式拉取，仅支持公共仓库。
        重复 attach 同一 PR（唯一键 issue_id+provider+repo_full_name+number）幂等返回既有记录并顺带刷新状态。attach 产生的行 `link_origin=manual`；PR 状态与 Issue workflow_status 不联动。
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachIssuePullRequestRequest'
      responses:
        '200':
          description: 关联记录（重复 attach 返回既有行并刷新状态）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IssuePullRequest'
        '400':
          description: 非 GitHub 或无法解析的 PR URL（40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          description: GitHub API 调用失败（50200；含 PR 不存在或无权访问）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/issues/{iid}/pull-requests/sync:
    post:
      tags: [issues]
      operationId: syncIssuePullRequests
      summary: 刷新该 Issue 下全部已关联 PR 的状态
      description: 逐条重新拉取 GitHub PR 信息并更新 state/merged_at/closed_at 等字段。只更新既有行，不新增、不删除；每条关联是独立失败域——单行拉取或保存失败记入 failures 不中断其余行（失败行保留旧数据），整体恒返回 200。仅 Issue/项目不存在返回 404，读取关联列表失败返回 500。
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        '200':
          description: 刷新结果：items 为刷新成功的关联行，failures 为逐行失败明细
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IssuePullRequestSyncResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/issues/{iid}/pull-requests/{id}:
    delete:
      tags: [issues]
      operationId: detachIssuePullRequest
      summary: 解除 Issue 与一个 PR 的关联（按关联行 id）
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
        - $ref: '#/components/parameters/PullRequestLinkId'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Issue 不存在（40405）或 PR 关联不存在（40412）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token，由 POST /api/auth/register、login 或 refresh 签发。仅 JWT 的端点对 `fsk_` API Key 返回 403（40301）。
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API Key，以 `fsk_` 开头（Web「设置 → API Key」创建）。与 jwtAuth 共用 `Authorization: Bearer` 头，服务端按 `fsk_` 前缀区分凭证类型。"
  parameters:
...
    IssueId:
      name: iid
      in: path
      required: true
      schema:
        type: string
      description: Issue ID（UUID，非 INT-123/GH-123 短编号）
    PullRequestLinkId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: PR 关联行 ID（UUID）
...
  responses:
    EmptySuccess:
      description: 操作成功（data 为 null）
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
...
    Unauthorized:
      description: 未提供或提供无效凭证（40100-40199）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: 资源不存在（40400-40499）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: 服务器内部错误（50000）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务错误码；0 表示成功
        message:
          type: string
          description: 人类可读信息；成功时为 "success"
        data:
          description: 响应负载；形状随端点而变，由 allOf 组合的兄弟 schema 收窄
      description: 统一 JSON 信封 {code, message, data}
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 业务错误码（4xxxx 按区间映射 HTTP 状态：401xx→401，403xx→403，404xx→404，409xx→409，412xx→412，50xxx→500；502xx→502）
        message:
          type: string
        data:
          type: object
          nullable: true
          description: 恒为 null
      description: 错误响应信封
...
    IssuePullRequestState:
      type: string
      enum: [open, closed, merged]
      description: merged = GitHub 上 closed 且 merged_at 非空
      x-go-type: model.IssuePullRequestState
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model
    IssuePullRequestLinkOrigin:
      type: string
      enum: [manual, synced]
      description: manual=用户显式 attach；synced=远端同步投影产生（清理逻辑只作用于 synced）
      x-go-type: model.IssuePullRequestLinkOrigin
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model
...
    AttachIssuePullRequestRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: GitHub PR 链接，形如 https://github.com/<owner>/<repo>/pull/<number>（允许 http、www 前缀与尾部 /files、query、fragment 等）；其他字段由服务端从 GitHub 拉取
...
    IssuePullRequest:
      type: object
      required: [id, issue_id, provider, repo_full_name, number, html_url, title, state, is_draft, author_login, head_ref, base_ref, merged_at, closed_at, link_origin, synced_at, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        issue_id:
          type: string
        provider:
          type: string
          description: 当前恒为 github；字段预留 gitlab
        repo_full_name:
          type: string
          description: owner/repo；允许与项目配置的仓库不同（跨仓库 attach）
        number:
          type: integer
        html_url:
          type: string
        title:
          type: string
        state:
          $ref: '#/components/schemas/IssuePullRequestState'
        is_draft:
          type: boolean
        author_login:
          type: string
        head_ref:
          type: string
        base_ref:
          type: string
        merged_at:
          type: string
          format: date-time
          nullable: true
          description: state=merged 时有值
        closed_at:
          type: string
          format: date-time
          nullable: true
        link_origin:
          $ref: '#/components/schemas/IssuePullRequestLinkOrigin'
        synced_at:
          type: string
          format: date-time
          description: 最近一次从 GitHub 刷新成功的时间
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    IssuePullRequestSyncFailure:
      type: object
      required: [id, error]
      properties:
        id:
          type: string
          description: 关联行 id（issue_pull_requests.id，detach 时用此 id）
        error:
          type: string
          description: 失败原因，前缀带 repo_full_name#number 便于定位
      description: PR 关联刷新失败明细；失败行保留旧数据
    IssuePullRequestSyncResult:
      type: object
      required: [items, failures]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IssuePullRequest'
          description: 本轮刷新成功的关联行（无关联或无成功时为空数组）
        failures:
          type: array
          items:
            $ref: '#/components/schemas/IssuePullRequestSyncFailure'
          description: 逐行失败明细
      description: syncIssuePullRequests 的响应形状，与批量 internal-meta 的 {items, failures} 约定一致
...